Move routines to casino.cpp/casino_card.cpp
