+++ casino.hpp
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace elona
+{
+
+void casino_dealer();
+void casino_choose_card();
+void label_1870();
+void label_1871();
+void label_1872();
+void casino_acquire_items();
+void casino_random_site();
+void label_1877();
+void casino_wrapper();
+bool casino_start();
+bool casino_blackjack();
+
+}
